Removed old KalmanFilterProcess
